fix #25 Allow to dispose LoopResources and PoolResources and listen
This commit adds a disposeAndListen method to both LoopResources and
PoolResources that allow to dispose these but also get a notification of
when the disposal is done, in a best effort fashion.

Subsequent calls however are not guaranteed to return a relevant Mono,
which could complete sooner than the actual disposal as triggered by the
first invocation.